FifaSearch/requirements.txt
PyPDF2~=2.7.0
tkcalendar~=1.6.1
python-dateutil~=2.8.2